[GECO-139] Addressing review comments
