+++ giles-eco/src/main/java/edu/asu/diging/gilesecosystem/web/core/kafka/KafkaProcessingListener.java
@@ -18,14 +18,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import edu.asu.diging.gilesecosystem.requests.ICompletedStorageDeletionRequest;
 import edu.asu.diging.gilesecosystem.requests.IRequest;
-import edu.asu.diging.gilesecosystem.requests.impl.CompletedStorageDeletionRequest;
 import edu.asu.diging.gilesecosystem.septemberutil.properties.MessageType;
 import edu.asu.diging.gilesecosystem.septemberutil.service.ISystemMessageHandler;
 import edu.asu.diging.gilesecosystem.web.core.service.IProcessingRequestService;
-import edu.asu.diging.gilesecosystem.web.core.service.core.ITransactionalDocumentService;
-import edu.asu.diging.gilesecosystem.web.core.service.delete.IDeleteDocumentService;
 import edu.asu.diging.gilesecosystem.web.core.service.processing.RequestProcessor;
 
 @PropertySource("classpath:/config.properties")
@@ -41,12 +37,6 @@ public class KafkaProcessingListener {
 
     @Autowired
     private ISystemMessageHandler messageHandler;
-
-    @Autowired
-    private IDeleteDocumentService deleteDocumentService;
-    
-    @Autowired
-    private ITransactionalDocumentService documentService;
     
     private ObjectMapper mapper;
 
@@ -84,17 +74,5 @@ public class KafkaProcessingListener {
 
         processingRequestService.addReceivedRequest(request);
         processor.handleRequest(request);
-    }
-
-    @KafkaListener(id = "giles.deletion.topic.listener", topics = "${topic_delete_storage_request_complete}")
-    public void receiveDeletionCompletedMessageMessage(String message, @Header(KafkaHeaders.RECEIVED_TOPIC) String topic) {
-        ICompletedStorageDeletionRequest request = null;
-        try {
-            request = mapper.readValue(message, CompletedStorageDeletionRequest.class);
-        } catch (IOException e) {
-            messageHandler.handleMessage("Could not unmarshall request.", e, MessageType.ERROR);
-            return;
-        }
-        deleteDocumentService.deleteDocumentAfterStorageDeletion(documentService.getDocument(request.getDocumentId()));
     }
 }
